Clean up helios-solo deployment via watchdog
If an integration test deploys containers using helios-solo
and the test runner dies without calling
HeliosDeploymentResource.after(), we could be left with stray containers
on the Docker Host.

Ensure we clean up by using a watch dog process that's spawned by the
solo/docker/start.sh script. This process will detect if the
HeliosSoloDeployment is no longer running (via the maintenance of a
TCP connection), and will undeploy stray jobs.