helios-solo: use bridge interface for http healthchecks
If the agent is using `unix://` to communicate with Docker, then the
`HttpHealthChecker` ends up using `localhost` as the host in the URL
it constructs to healthcheck the job.

When the agent is running in helios-solo (i.e. in a Docker container
itself), the agent uses the default `unix:///var/run/docker.sock`
endpoint to talk to Docker but a HTTP request to
`localhost:external-port` will not work as `localhost` refers to the
container itself.

In the scenario where the agent is running in helios-solo *and*
docker.host is using the unix protocol, this change will have the
`HttpHealthChecker` instead use the bridge address of the container to
send it's healthcheck to.

In order for the agent to know if it is running within helios-solo,
added a system property in the helios-solo startup script. This property
is also enabled for the master as well in case it is needed in the
future.