fix occasional NullPointerException in DefaultDeployer
For currently unknown reasons, occasionally
`client.hostStatus(candidateHost)` will return a completed
ListenableFuture with a `null` value when called by DefaultDeployer
(which HeliosClient does when the host doesn't exist). It's not clear
why the host would have `null` status when a few lines earlier the host
was retrieved from `client.hosts()`.

But in any case, this fixes the NPE that occurs when this happens.