implementation of FastForward metrics reporter
Reports metrics from the v3 MetricsRegistry to [ffwd][].

The String "name" of the Metric is sent as the 'what' attribute in the
FastForward metric. No translation of the (possibly long, Java
classname-ish) name is done.

The type of metric is included in the 'metric_type' attribute. For
meters, timers, and histograms, the various component values of the
com.codahale.metrics.Metric (m1/m5/m15, p50, p99 etc) are sent as
individual Metric instances. The 'stat' attribute is filled out with the
corresponding statistic.

Note that not every statistic from Meters/Timers/Histograms is sent to
FastForward in order to limit the amount of data being sent and needing
to be stored downstream. For Metered instances, only 1m and 5m is sent.
For Histograms, just median, p75 and p99 (plus min/max/mean/stddev).

[ffwd]: https://github.com/spotify/ffwd