client: Cache generated X509 certificate
If we generate an X509 certificate and get it signed by ssh-agent, cache
the resulting certificate and key in ~/.helios. This can reduce the time it
takes for future requests by around a second since generating new key pairs
is an expensive operation.

To do this, `X509CertificateFactory` now uses `CertificateAndPrivateKey`
and we ditch the basically equivalent `CertificateAndKeyPair` class. The
logic for reading X509 certificates/keys from files has been moved from
`CertificateFileHttpsHandler` to `CertificateAndPrivateKey`, where it can
be shared by `CertificateFileHttpsHandler` and `X509CertificateFactory`.