remove race conditions from FastForwardReporterTest
Previously the test would start the FastForwardReporter and it's use of
a `ScheduledExecutorService` to report metrics every 10 milliseconds.

This seems to cause some strange effects when used with Mockito's
`timeout()` method, as the latter tries to run all of the verifications
against the mock object over and over again until the `timeoutMillis`
value has elapsed.

In addition, it seems as if the `report` method cannot reliably be
executed at a rate of 10 milliseconds; the value seems too small.

It is much simpler for the test to force the FastForwardReporter to
report it's metrics once on the main thread so that the Test doesn't
have to worry about concurrency or timeouts or scheduling.