HeliosClient: catch EndpointIterator with no endpoints
Handle the case where the HeliosClient is configured with a list of URIs
which resolve into no actual Endpoints when EndpointIterator does it's
work.

Currently if the URI resolves into no endpoints (if there are DNS
issues, or the URI just can't be resolved), the error is only caught
when the request is made in RetryingRequestDispatcher because
EndpointIterator.next() will throw a `java.util.NoSuchElementException`,
and the RetryingRequestDispatcher will retry the doomed request many
times.