refactor ConfigTest so it passes in any environment
Current test relies on resolving internal spotify.net addresses, and
causes exceptions with the previous commit that throws exceptions when
the HeliosClient is constructed if there are no endpoints to connect to.

This refactoring makes sure that `HeliosClient.Builder.build()` is not
actually called within the test, and even better that the test passes on
any network (or without a network configuration).