allow alternate java installation in startup scripts
Changes `bin/helios-agent` and `bin/helios-master` to respect a `$JAVA`
environment variable which contains the path to the java command to run
the helios process with.

If the environment variable is not set, then the scripts fall back to
(the prior behavior of) a simple `java` command.

This can be useful for environments where the helios agent and/or master
should be run with a version of Java that is not the default version on
the host.