do not signal agent registration until registration complete
Currently if the AgentZooKeeperRegistrar cannot register the agent
because of a hostname conflict, it attempts to shut down the
AgentService and exits itself; however, the ZooKeeperRegistrarService
counts down on the "registration signal" anyway.

Change the registrar behavior so that a hostname collision causes
AgentZooKeeperRegistrar to return it's failure, so that the
ZooKeeperRegistrarService backs off and retries later, as it would for
Zookeeper exceptions. The registration signal is only counted-down on an
actual successful registration, preventing possible race conditions
with any threads that might be waiting on that signal to do work in
Zookeeper that should only proceed when registration is successful.

This causes the agent registration to be retried later, according to
ZooKeeperRegistrarService's retry policy, since the AgentService (and
the agent itself) is no longer terminated.