Fix race btw HostInfoReporter + ZKRegistrarService
Wait for the agent to register itself with ZooKeeper to
prevent HostInfoReporter from winning a race in updating the
ZooKeeper node /status/hosts/<hostname>/hostinfo.

The ZooKeeperRegistrarService run by the agent checks the mtime
of the above znode and only tries to re-register if that znode
has been stale for a while (defaults to 10 minutes).

If we don't make HostInfoReporter wait, the agent will never
re-register itself in ZK.