block HostInfoReporter and AgentInfoReporter until agent is registered
Currently there is a race condition that occurs when an agent with a
hostname collision (it has the hostname of an agent already registered
in Zookeeper but it's `id` does not match) attempts to register itself.

If the automatic-deregistration logic does not kick in,
AgentZooKeeperRegistrar attempts to shut down the AgentService and
therefore the agent itself.

However the HostInfoReporter and AgentInfoReporter have already been
started asynchronously by the AgentService. The HostInfoReporter
in particular overwrites the existing data at
`/status/hosts/<hostname>/hostinfo` with the HostInfo of the hostname
with the collision (and which is about to be shut down).

Overwriting the `/status/hosts/<hostname>/hostinfo` path causes it's
`mtime` to be updated. This is the same exact `mtime` that
AgentZooKeeperRegistrar uses to decide if enough time has passed to do
the automatic deregistration of the old node.

This can cause the new agent (for instance, after reinstalling the
operating system) to get stuck in a loop where it tries to register,
fails, shuts down, and is restarted by Upstart. This loop will never end
because of the race condition that will bump the mtime of the Znode
(which seems to win the race at least once every few iterations).

To fix this, introduce a new type of InterruptingScheduledService that
awaits on a CountDownLatch before executing it's task. This class is
then extended by HostInfoReporter, LabelReporter, etc.