avoid use of MoreObjects.toStringHelper
Users of `helios-testing` that are stuck on previous versions of Guava
(for example, if building something that needs to use Hadoop, and using
TemporaryJobs within the test of that thing) fail to load `MoreObjects`
from the classpath as this was only introduced in Guava v18.

Since the functionality used here to implement `toString()` is so tiny,
it feels unnecessary to rely on an external library for it (which
creates the potential for conflicting dependency library problems).
Therefore just implement `toString()` with simple String concatenation
(which the compiler will replaced with StringBuilder for us).