support dockerfile-maven-plugin in TemporaryJobsBuilder.imageFromBuild()
Currently, the `imageFromBuild()` method in `TemporaryJobsBuilder()`
supports reading the `image_info.json` produced by docker-maven-plugin
to find the name of the image to use in the TemporaryJob. This file is
read from either the classpath or the filesystem if the former fails.

This change adds support for reading the `target/docker/image-name` file
produced by dockerfile-maven-plugin as well.