fix SimpleTest relying on build output from other projects
This test is using `imageFromBuild()` to create a TemporaryJob,
seemingly relying on the docker-maven-plugin output from some other
module in the Maven project. The pom.xml in helios-testing does not
create any images on it's own.

I'm not sure how this test worked before; but since it doesn't seem to
actually care what image it uses, changed it to specify `busybox`.