Introduce a 'rolling operation' concept.
The target job ID and rollout options of a rolling update are associated with
the rolling operation - not the deployment group itself. Each rolling
operation is associated with a deployment group. The 'reason' for a rolling
update (i.e. manual or 'hosts changed') is a property of the rolling operation.

Deployment groups may now have state 'stable' or 'unstable'. Stable deployment
groups are known to be running the same job across all agents. Unstable groups
are known to have some containers down, or to be running different jobs across
agents. Rolling operations model the states previously associated with
deployment groups (ROLLING_OUT, DONE, FAILED).

Some hacks (i.e. the DeploymentGroupResponse) have been put in place in order
to avoid breaking the existing REST and CLI APIs.

This change means:
* A failed rolling update will no longer require human intervention to unblock
  future automatic (i.e. 'hosts changed') rolling updates.
* Rolling restarts could be implemented as a variant of a rolling operation -
  perhaps as a rolling operation with a null jobId?
* We could expose an audit trail of a deployment group's rolling updates.