[helios-testing]: Ensure helios-solo agent is UP
in addition to registering with ZooKeeper.

This should prevent TemporaryJobs failures where `DefaultDeployer.deploy()`
gets a list of hosts and then tries to deploy to the single solo agent
only to find it's not UP yet.

This is a race condition where the solo agent finishes the first part of
`AgentZooKeeperRegistrar.tryToRegister()` where it writes data to ZK's
/config/host/<hostname>/id which makes it appear in calls to
`HeliosClient.listHosts()`, but hasn't yet written the UP node
to ZK's /status/host/<hostname>/up in the
latter part of `tryToRegister()`.