TemporaryJobs: handle DOCKER_HOST with unix sockets
`TemporaryJobs.builder()` attempts to set up default values for the
`endpoints` to connect to (representing the address of the Helios
cluster to talk to), prior to the user setting any endpoints or
HeliosClient of their own.

After checking a few environment variables for which Helios host to talk
to, `TemporaryJobs.Builder` then checks if `DOCKER_HOST` is set and uses
it as the host in a `http://` URI.

This fails if `DOCKER_HOST` is a URI to a Unix socket (e.g.
`unix:///var/run/docker.sock`), such as when using Docker for Mac.

To fix this, change the Builder to ignore `DOCKER_HOST` values that
start with `unix://`, and rely on the user to set the endpoints or
HeliosClient to use themselves.