helios-testing: Add ExistingHeliosDeployment
User can pass `ExistingHeliosDeployment` as a parameter to `TemporaryJobs`'
`HeliosDeployment.  `HeliosSoloDeployment` won't be used in this case.

Create the package-local interface `Undeployer` and implement it with
`DefaultUndeployer`. The interface is needed to easily unit-test
`ExistingHeliosDeployment`. The `TemporaryJob.undeploy()` method is
still the only publicly accessible method used to undeploy a
`TemporaryJob`.

I also write a test for `DefaultUndeployer`.

I comment out a test in `ConfigTest` because I couldn't figure out a
nice way to make it pass given these new changes. I started adding a way
to pass `HeliosClient.Builder` to `TemporaryJobs`, and it quickly became
horrible...

Signed-off-by: David Xia <dxia@spotify.com>