helios-testing: Clean up if ExistingHeliosDeployment
Add a flag to `TemporaryJobs` (is there a better way than this?)
indicating if the `HeliosDeployment` is an `ExistingHeliosDeployment`.

If it is, undeploy jobs at the end and remove the job prefix file and
close the resource. Otherwise skip all this since helios-solo's
`start.sh` contains all this logic already.

Have `ExistingHeliosDeployment` close its `HeliosClient` if it created
its own instance and not to close when one has been passed in.

Update some tests to use `@ClassRule` for `HeliosDeployment` instead of
`@Rule` to reduce overhead.

Signed-off-by: David Xia <dxia@spotify.com>