Don't block rolling updates when new hosts fail to update.
Currently it's possible that:
1. A user creates a deployment group and successfully performs a rolling update
2. New hosts are added to said group, triggering another rolling update
3. One or more of those hosts fails to update, and the group is marked FAILED
4. More hosts are added, but we don't trigger a rolling update because we're wary the last one failed.

This change prevents the scenario outlined at step 4 above, while ensuring we
don't trigger an automatic rolling update for new hosts if the most recent
manual rolling update failed.

This is implemented by tracking the reason (manual, or because of changed hosts)
of the last rolling update in the DeploymentGroup object. This gives us an extra
data point when deciding whether to automatically rolling update upon the
addition of new hosts to a deployment group.