helios-env: fix for Docker for Mac
When `DOCKER_HOST` is not set, the script assumes that if the local OS
is Linux then the Unix socket at `/var/run/docker.sock` can be used to
communicate with the Docker daemon.

Docker for Mac also uses this socket and path and does not need
`DOCKER_HOST` to be set to work, so add a test to see if a bare `docker
info` command exits with a success and use the same Unix socket path
when it exits successfully.