always update status to STOPPING when stopping a job
Supervisor stores a status update of STOPPED after the container has
been shut down. For jobs with grace periods, it always stores a status
update of STOPPING prior to unregistering the job from service discovery
and sleeping before the container is stopped.

It would make more sense to unconditionally store the status update of
STOPPING prior to the container stop request and the STOPPED state,
since stopping the container may take a number of seconds for the
container to clean itself up or otherwise respond to SIGTERM. This would
lead to more consistent output from the `helios history` command.

Closes #936.