Add JobHistoryReaper
This reaper, disabled by default, runs in the master and will
remove job histories whose corresponding jobs no longer exist.
It runs once a day when enabled.

This reaper is meant to address the scenario where jobs are deleted but
their histories are left behind in ZooKeeper. See JobHistoryReaper's
JavaDocs for more details.