replace "labels" with "selectors" for clarity
When filtering the list of hosts, refer to the argument as "selectors"
rather than "labels" as the latter term is used with a different meaning
within the helios-agent and helios-master.

Also change the argument from List to Set to reflect that the collection
of selectors does not have any real ordering.

In addition, changed HostListCommand argument `--labels` to be
`--selector` for the same reason.