fix NPE in Supervisor.Stop when job has gracePeriod
It's possible for a Supervisor to receive a goal of UNDEPLOY for a job
that it has not been told to START - perhaps because the agent was
restarted during the undeploy of a previous job (during the grace period
sleeping).

Since the TaskRunner is only set after receiving the START goal, the
call to `runner.unregister()` when the job being undeployed throws a
NullPointerException.

Add a test to verify behavior.