HeliosSoloLogService: use timeouts with Futures
In a test class that uses HeliosSoloDeployment as a `@Rule` (i.e. one
helios-solo instance per test method), I found that the test was hanging
for a long period of time after the test itself seemed to be finished
(judging from the class-under-test's log output).

Thread dumps showed that the thread running HeliosSoloLogService was
blocked at a call to `heliosClient.listHosts().get()`. This change adds
timeouts when using `Future.get()` to avoid any possibility of blocking
forever.

Also change the Service itself to use daemon threads with it's executor,
as we do not wish to keep the JVM running under any instance if the
HeliosSoloLogService is still running when the main thread has exited.