add logically-equal comparison test for HostSelectors
This adds a test to compare HostSelectors and Collections of
HostSelectors to find out if they are "logically equal". Currently the
only logically equal selectors are `foo = bar` and `foo in (bar)`.

This is intended to  remove an error message about conflicting host
selectors if a user tries to create a deployment group with selectors
`[foo in (bar), pool = a]` when a deployment group already exists
with selectors `[foo = bar, pool = a]`.