Deploy to all hosts on hosts changed, not just new ones.
This reintroduces our original strategy of generating rolling update operations
for all the current hosts of a deployment group, even if they're already running
the job we're deploying.

This is necessary to avoid a race where:
- HOSTS_CHANGED update A determines set X of new hosts, and writes update ops
- HOSTS_CHANGED update B determines set Y of new hosts, and writes update ops

Rolling update B's ZK ops could overwrite update A's before it had finished
rolling out, meaning some of set X of hosts would not be deployed to until a
MANUAL rolling update was requested.