Do the create-then-delete dance for all nodes when deleting a DG
For each new deployment group ZK path we add (statusDeploymentGroupRemovedHosts
in this instance) while Helios is in production we must first ensure a DG has
that path before trying to delete it.

Instead of adding a bespoke exists() check around the delete() for the newly
added path I've applied the statusDeploymentGroupTasks 'safe delete' pattern for
all paths.

One implication of this is that calls to delete a non-existent deployment group
will now succeed rather than throwing DeploymentGroupDoesNotExistException. Unix
style!