periodically check connectivity to Cloud Pubsub
when the GooglePubSubSender cannot connect to the pubsub service,
disable message publishing until connectivity is re-established (at the
next scheduled time).

Along the way, I refactored the common EventSender-instantiation code
from AgentService and MasterService into a new class named EventSenderFactory.
To make this common class possible, also extracted a class for the
common configuration fields between AgentConfig and MasterConfig. This
new CommonConfiguration class is incomplete - more fields can be moved
to this class to avoid duplication, but I left that for future commits.