UpdateDeploymentGroupHosts only uses UP hosts
When there are a lot of DOWN hosts, `UpdateDeploymentGroupHosts`
makes a lot of unnecessary ZK reads via `MasterModel.getHostLabels() for
hosts that are DOWN.

This PR might not reduce ZK ops. `listUpHosts()` is 2n reads, where n is
the number of hosts. `listHosts() + `getHostLabels()` is n + 2n = 3n.

If we do the TODO in ZKMM to only list hosts that have id nodes, aka
completed registration, that makes `listUpHosts()` 3n. In cases,
where there aren't many DOWN hosts, this increases ZK ops.

It only helps in cases where there are many DOWN hosts.