monitor how often zookeeper connection state changes
Adds meters for how often the state of the zookeeper connection changes.

ReportingZooKeeperClient sets a listener on the underlying connection
which calls a method in ZooKeeperModelReporter which in turn notifies
the ZooKeeperMetrics instance.

Within the ZooKeeperMetrics implementation, one meter is kept for the
rate of overall connection state changes, and a set of meters is kept
for each individual ConnectionState value, to track how often each state
is being hit.