More robust cleanup of removed hosts for deployment groups
Instead of the previous method where we would try to undeploy from hosts
that were no longer registered, instead just apply a straightforward filter
to the removed hosts list.

This is clearer and works without generating a bunch of unnecessary tasks.
Also, the previous version would only work if the set of UP hosts for the
deployment group changed. This version doesn't have that flaw.