add maven-enforcer-plugin rule to ban netty-* <= 4.1.3
Enforces that any netty dependencies with the memory leak (see commit
6c7e030) are not used/included.

Also moves the maven-enforcer-plugin configuration from the
pluginManagement section of the parent pom to the plugins section, so
that it actually runs and applies to each submodule.

Now that maven-enforcer-plugin is running for each submodule as
intended, we have a few requireUpperBoundDeps violations that have to be
fixed as well.