remove fallback for batch status fetching
When the bulk jobStatus method was first added to the API (back in
commit b1d345f), a fallback was added in case the bulk query failed to
loop over the list of jobs and fetch the status of each individually.

Assuming this fallback was added to handle having a newer CLI version
talking to a version of the master that didn't have this API yet, I
think it is safe to remove this logic at this point.