JobListTest: don't ignore json parsing errors
If the test encounters a JsonParseException when trying to deserialize
the json response from the `helios hosts` CLI command, don't ignore the
exception and continue the test.

It seems like we would want to know if the command was asked to return
json output but failed to do so.

I would guess that this ignore-some-exceptions logic was a result of the
JobListCommand sometimes returning an empty json array rather than an
empty json object, which was fixed in commit 3b71ade.