TemporaryJobs: move old job cleanup to apply()
Instead of attempting to clean up new jobs in the TemporaryJobs
constructor, instead perform the cleanup (just one time) when the
Statement is evaluated.

In other words, we attempt to cleanup old jobs when the test method is
being executed (if the user is using `@Rule`).

This avoids conflicts when used with HeliosSoloDeployment in a RuleChain
(as the docs suggest) like:

```
@ClassRule
public static final RuleChain chain = RuleChain
    .outerRule(soloResource)
    .around(temporaryJobs);
```

If the old job cleanup is performed in the constructor, then
TemporaryJobs will start attempting to query the Helios cluster before
it has been fully initialized by the HeliosDeploymentResource named
`soloResource` in the above example.