fix image repository name validation
It seems like at some point Docker changed the format of image names to
be less restrictive: the image name can have multiple slash-separated
name components (with no bound on the number of components):

> The image name which is made up of slash-separated name components...
> Name components may contain lowercase characters, digits and
> separators. A separator is defined as a period, one or two
> underscores, or one or more dashes. A name component may not start or
> end with a separator.

This commits removes the regexes in JobValidator that check the
"namespace" and "repoName" separately along with the limits on how long
each portion can be. It now validates that each slash-seperated name
component contains only the allowed characters and that the entire
repository name is below the limit of 255 characters.

See:
https://github.com/docker/docker/commit/ea98cf74aad3c2633268d5a0b8a2f80b331ddc0b
https://github.com/docker/distribution/blob/master/reference/reference.go
https://github.com/docker/distribution/blob/master/reference/regexp.go

Fixes #1070.