don't assign default values to PortMapping.ip
We can't assign default values from within the POJO class to any
"descriptors" that are reachable from the Job class. When the Job
builder computes the JobId.hash, it takes all of the Job parameters that
are not meta parameters (username, create time), serializes them to
JSON, and then computes a sha1 hash for that string.

If we change things so that a client running an older version of the
code does not have the same default values for fields that the server
does, then the JobId.hash computed on the server-side will be different
than what the client computed.

One way that this will surface itself is via users of helios-testing and
TemporaryJobs; since HeliosSoloDeployment always pulls
`spotify/helios-solo:latest`, regardless of what version of
helios-testing is being used.

Another possible way this problem would express itself is if an older
helios-cli user tried to create a job against a newer helios master.