simplify ZooKeeperHealthChecker by removing PathChildrenCache
It is not clear why this healthcheck creates and starts a
PathChildrenCache for the purposes of knowing if zookeeper is conneted
or not - the PathChildrenCache proactively attempts to keep all children
of a ZK path locally cached. This involves registering watches on those
paths, responding to the events, keeping the data in-memory, etc.

Keeping the /status/hosts tree cached on the masters would be a pretty
heavyweight operation - the cache kept by the healthcheck is updated
every single time any host in the cluster has a status change, has a job
deployed, etc.

Since the only thing that this healthcheck really cares about is the
state of the zookeeper *connection* itself, and not the state of a tree,
replace this with a ConnectionStateListener.