helios-solo: skydns workaround to allow DNS responses up to 32768 bytes
TL;DR; skydns does not handle TCP well. We already worked around this
in https://github.com/spotify/helios/pull/900. See thar PR for more context.
However, that change only fixed the problem to an extent as we still have
the same issue once the responses are >4096 bytes.

This change extends that workaround to allow us to survive responses up to
32768 bytes in size. This change does not fix the issue, but should make it
more rare.