master: add 'hostPattern' parameter to GET /jobs
Add the ability to filter on the server-side the output from `GET /jobs`
to jobs deployed to specific hosts.

This is intended to improve the performance of the
`helios status --host` command, which currently runs throught the
following steps on the client-side in order to output the status of jobs
deployed on hosts that match the `--host` name pattern:

1. Get all jobs in the cluster (GET /jobs)
2. Get the job status of all these jobs (POST /statuses)
3. Filter the response to jobs deployed on hosts that match the filter

In a cluster with thousands of hosts, this is quite inefficient and
results in a lot of data being transferred that the CLI ends up
discarding. As an example, in a cluster with ~3000 jobs, the first
request ends up transferring more than 3MB of JSON, while the second
request transfers 14MB of JSON.

This can be improved by moving the filtering onto the server-side, so
that the CLI can simply ask for jobs that are deployed on a given host
name pattern in step 1 above. This also allows the server to not bother
to make thousands of zookeeper calls when finding the status of all
jobs.