Upgrade logback for faster CLI startup
logback calls `InetAddress.getLocalHost()`, which is painfully slow on
some versions of macOS (i.e., 4-5 seconds). Since this happens when
logging is being configured, it potentially adds seconds to every
`helios` CLI invocation.

To address this, we upgrade to a newer version of logback that includes
a fix to evaluate the hostname lazily and not block startup:
https://jira.qos.ch/browse/LOGBACK-1221