skip AWAIT_RUNNING in rollout plan if ignoreFailures is true
Do not add AWAIT_RUNNING to the list of rollout tasks if
options.ignoreFailures is true.

This will have the effect of a rolling-update with this option set
creating a list of tasks that does an UNDEPLOY, DEPLOY on each host in
the group, without waiting for the job to finish
pulling/running/healthchecking on each host.