do not set DeploymentGroup to FAILED when ignoreFailure is true
When rolloutOptions.ignoreFailures is true, the RollingUpdateOpFactory
will quietly transition to the next task in the deployment group's
rollout plan (the DeploymentGroupTasks object) when one of the overlaods
for `error(..)` is called.

This will have the effect of moving the rolling update to the next task
in the plan when the group/rollout otherwise would have FAILED, only if
the ignoreFailures option is set to true.

At first glance one problem with this method of internally returning the
result of `nextTask()` when `error(..)` is that sometimes `nextTasks` is
called with a list of ZookeeperOperations supplied as a parameter by the
calling code; however, this only occurs when the ZookeeperMasterModel is
first processing an action of DEPLOY, MARK_UNDEPLOYED or UNDEPLOYED. The
overload of `nextTasks()` with no argument is called whenever the task
action is AWAIT_RUNNING and the task reaches running, therefore it
should be safe (unless the callers are refactored).

This is an alternate approach to #1103 to try to achieve the same goals
but with less downside.