add a metric that reports total agent health
When monitoring Helios agents internally at Spotify, we query our
metrics systems for all metrics in
`[helios.docker.ok, helios.dockerd.ok, helios.deadlocks.ok]` and then do
an aggregation that is `min value by host`.

Because this needs to return 3 time series for every agent we have in
our backend, the query hits an upper limit of how many time series the
monitoring system can return.

To be able to have more efficient queries, and to avoid the monitoring
queries from having to keep track of which metric names are the ones to
monitor for "health", this change introduces a new metric named
`helios.healthy` that reports a 1 if all healthchecks are OK, and a 0 if
any single healthcheck (or all) fails.