[solo] Disable unbound using IPv6; we don't need it
A [bug in Docker 17.04][docker-bug] disabled IPv6 causing
helios-solo to crash when trying to run unbound. Unbound tries to
open a control/admin port on localhost 8953 and defaults to listening
on both IPv4 and IPv6.

This change forces unbound to listen **only on IPv4 for all ports**.

Strange things:
  * unbound docs say remote-control->control-enable defaults to no
  * for some reason unbound can listen to port 53 with IPv6

fixes #1116

  [docker-bug]: https://github.com/moby/moby/issues/32433