integrate with Google ContainerRegistry support in docker-client
Updates docker-client to 8.6.0 which has support for configuring the
DockerClient instance to authenticate against Google Container Registry.

The credentials to use in authenticating for pulls from GCR are
configured with a new flag `--docker-gcp-account-credentials`. The GCR
authentication layer is only enabled when this flag is set. This
argument value is expected to be a file containing the JSON key for the
service or user account to pull images with.

Also adds support for a flag
`--docker-use-gcp-application-default-credentials` to use the
[Application Default Credentials][0] with GCR.

[0]: https://developers.google.com/identity/protocols/application-default-credentials